+++ config/routes.rb
@@ -2,6 +2,7 @@
 
 Markdowndocs::Engine.routes.draw do
   root "docs#index"
+  get "search_index", to: "docs#search_index", as: :search_index
   get ":slug", to: "docs#show", as: :doc
   resource :preference, only: [:update]
 end