+++ lib/markdowndocs/configuration.rb
@@ -4,7 +4,7 @@ module Markdowndocs
   class Configuration
     attr_accessor :docs_path, :categories, :modes, :default_mode,
       :markdown_options, :rouge_theme, :cache_expiry,
-      :user_mode_resolver, :user_mode_saver
+      :user_mode_resolver, :user_mode_saver, :search_enabled
 
     def initialize
       @docs_path = nil # Resolved lazily so Rails.root is available
@@ -16,6 +16,7 @@ module Markdowndocs
       @cache_expiry = 1.hour
       @user_mode_resolver = nil
       @user_mode_saver = nil
+      @search_enabled = false
     end
 
     # Lazily resolve docs_path so Rails.root is available